Added PUT/POST/GET/DELETE to mirror occi-SO behavior
Outlined the functionality for all 4. Requires a pre-configured
nfvo and the to-be-deployed NSD-id in application.properties. This
should fallback to uploading vnfds and NSD in the future.

GET is malformated and only returns the private IP's of the components,
this will also contain floating IP's and the proper mcn-namingscheme
in the future.